[US240431] - Added unit tests for connecting to public broker.
